# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the Machine Controller API.",
        "title": "Machine Controller API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Elizandro Dantas",
            "url": "https://www.linkedin.com/in/elizandrodantas",
            "email": "dantaspm@icloud.com"
        },
        "version": "2.0"
    },
    "paths": {
...
        "/machine": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "List machines",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Machine"
                ],
                "summary": "List machines",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by machine ID",
                        "name": "machine_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by OS",
                        "name": "os",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter with query",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.MachineListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.HttpError"
                        }
                    }
                }
            },
            "post": {
                "description": "Machine registration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Machine"
                ],
                "summary": "Machine registration",
                "parameters": [
                    {
                        "description": "MachineData",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.MachineData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.MachineResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.HttpError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/domain.MachineResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.HttpError"
                        }
                    }
                }
            }
        },
        "/machine/update-name": {
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Update machine name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Machine"
                ],
                "summary": "Update machine name",
                "parameters": [
                    {
                        "description": "MachineUpdateNameRequest",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.MachineUpdateNameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.HttpError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.HttpError"
                        }
                    }
                }
            }
        },
        "/machine/{id}": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get machine detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Machine"
                ],
                "summary": "Get machine detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Machine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Machine"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.HttpError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.HttpError"
                        }
                    }
                }
            }
        },
        "/notes/create": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Create a note",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notes"
                ],
                "summary": "Create a note",
                "parameters": [
                    {
                        "description": "NotesCreateRequest",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.NotesCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.HttpError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.HttpError"
                        }
                    }
                }
            }
        },
...
        "/rule/actives/{machine_id}": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get actives rules",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rule"
                ],
                "summary": "Get actives rules",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Machine ID",
                        "name": "machine_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.RuleActivesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.HttpError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.HttpError"
                        }
                    }
                }
            }
        },
        "/rule/create": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Create a rule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rule"
                ],
                "summary": "Create a rule",
                "parameters": [
                    {
                        "description": "RuleCreateRequest",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.RuleCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.RuleCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.HttpError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.HttpError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.HttpError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "domain.Machine": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "guid": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "os": {
                    "type": "string"
                },
                "query": {
                    "type": "string"
                }
            }
        },
        "domain.MachineData": {
            "type": "object",
            "required": [
                "expire",
                "guid",
                "name",
                "os",
                "service_id"
            ],
            "properties": {
                "expire": {
                    "type": "integer"
                },
                "guid": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100
                },
                "os": {
                    "type": "string",
                    "maxLength": 100
                },
                "service_id": {
                    "type": "string"
                }
            }
        },
        "domain.MachineJson": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "guid": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "os": {
                    "type": "string"
                },
                "query": {
                    "type": "string"
                }
            }
        },
        "domain.MachineListResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.MachineJson"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "domain.MachineResponse": {
            "type": "object",
            "properties": {
                "identify": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "service_identify": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "domain.MachineUpdateNameRequest": {
            "type": "object",
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
        "domain.NotesCreateRequest": {
            "type": "object",
            "required": [
                "machine_id",
                "text"
            ],
            "properties": {
                "machine_id": {
                    "type": "string"
                },
                "text": {
                    "type": "string",
                    "maxLength": 5000,
                    "minLength": 1
                }
            }
        },
...
        "domain.RuleActivesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.RuleJoinServiceJson"
                    }
                },
                "machine_id": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "domain.RuleCreateRequest": {
            "type": "object",
            "required": [
                "expire",
                "machine_id",
                "service_id"
            ],
            "properties": {
                "expire": {
                    "type": "integer",
                    "minimum": 0
                },
                "machine_id": {
                    "type": "string"
                },
                "service_id": {
                    "type": "string"
                }
            }
        },
        "domain.RuleCreateResponse": {
            "type": "object",
            "properties": {
                "expire": {
                    "type": "integer"
                },
                "machine_id": {
                    "type": "string"
                },
                "service_id": {
                    "type": "string"
                }
            }
        },
...
        "domain.RuleJoinServiceJson": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "expire": {
                    "type": "integer"
                },
                "machine_id": {
                    "type": "string"
                },
                "service_id": {
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}